backend/requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-unstructured
unstructured
faiss-cpu
pydantic
python-dotenv
requests
langchain-groq
langchain_groq
langchain
chromadb
numpy
motor
python-jose
passlib
sklearn
scipy
motor